+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `Beaker.experiment.latest_job()` method.
+- Added `strict: bool` argument to `Beaker.(experiment|job).(wait_for|as_completed)` methods.
 
 ## [v0.15.1](https://github.com/allenai/beaker-py/releases/tag/v0.15.1) - 2022-05-06